Interceptors are skipped for mocked activities
If you mock an activity and you define return values, the interceptors are
skipped. It is not consistent with the local activities which run interceptors.
This commit simply wraps the return values into a mock function and it is
passed to activity/workflow executor which internally calls all interceptors.

Test plan: added unit tests